Se cambia la vista de Autenticar, se quita el Emoji y se muestra un CheckMark. || Se cambia el Double de Amount en Transaction por Decimal.

+++ MySpend/MySpend/Presentation/Scenes/Main/Resume/Resume.swift
@@ -10,5 +10,5 @@ import Foundation
 struct Resume {
     var userName: String = ""
     var transactions: [TransactionModel] = []
-    var totalBalance: Double = 0
+    var totalBalance: Decimal = 0
 }
